% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupriseMeFeature.R
\name{addSurpriseMeObserver}
\alias{addSurpriseMeObserver}
\title{Suprise me function}
\usage{
addSurpriseMeObserver(input, output, session, cocktails, selected_cocktail)
}
\arguments{
\item{input}{Shiny input vector}

\item{output}{Shiny output vector}

\item{session}{Shiny session}

\item{cocktails}{cocktails dataframe}

\item{selected_cocktail}{Selected row of specific cocktail}
}
\value{
Shows Cocktail Details page with random cocktail
}
\description{
This function detects when the suprise me button is clicked and assigns a random
cocktail to selected_cocktail()
}
