% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocktailApp.R
\name{cocktailApp}
\alias{cocktailApp}
\title{Shiny app function}
\usage{
cocktailApp()
}
\value{
Laucnhes the shiny app
}
\description{
The \code{cocktailApp} function launches the shiny app directly from the package.
}
\examples{
\dontrun{cocktailApp()}

}
