% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertUnits.R
\name{convertUnits}
\alias{convertUnits}
\title{Measurement Units Conversion (volume, weight and length)}
\usage{
convertUnits(data, conversion_direction)
}
\arguments{
\item{data}{A dataframe}

\item{conversion_direction}{A string to define conversion direction
(imperial_to_international or international_to_imperial)}
}
\value{
A new dataframe convert to either imperial or international units
}
\description{
The \code{convertUnits} function is used to convert imperial units to international, and vice versa.
It takes as input a data frame which contains values and units in volume, weight and height,
and converts those units to the desired system.
}
\examples{
\dontrun{convertUnits(cocktails, "imperial_to_international")}
}
