% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterCocktails.R
\name{filterCocktails}
\alias{filterCocktails}
\title{Filtering cocktails list}
\usage{
filterCocktails(df, alcohol = NULL, ing1 = NULL, ing2 = NULL)
}
\arguments{
\item{df}{The cocktails dataframe included in the package}

\item{alcohol}{A string which is in the alcohol_Vec vector}

\item{ing1}{A string which is in the side_ing_vec vector}

\item{ing2}{A string which is in the side_ing_vec vector}
}
\value{
A filtered drinks dataframe based on inputs
}
\description{
The \code{filterCocktails} function is used to create the list of
drinks based on inputs from the dropdowns.
It filters the datasets based on one alcohol and two side ingredients
and returns a filtered dataframe with drinks corresponding to those criteria.
}
\examples{
filterCocktails(cocktails, "gin")
filterCocktails(cocktails, "vodka", "cranberry juice")
}
