% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filtering}
\alias{filtering}
\title{Filtering cocktails}
\usage{
filtering(x)
}
\arguments{
\item{x}{Input ingredients}
}
\value{
filtered dataframe
}
\description{
This function filters cocktails from the complete
list based on the ingredients that were input by the user.
}
\examples{
  x
}
