% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering_fun.R
\name{filtering_fun}
\alias{filtering_fun}
\title{Filtering for cocktails list}
\usage{
filtering_fun(df, alcohol = NULL, ing1 = NULL, ing2 = NULL)
}
\arguments{
\item{df}{A dataframe}

\item{alcohol}{A string}

\item{ing1}{A string}

\item{ing2}{A string}
}
\value{
a filtered drinks dataframe based on inputs
}
\description{
The `filtering_fun` function is used to create the list of
drinks based on inputs from the dropdowns.
It filters the datasets based on on alcohol and two side ingredients
an returns a filtered dataframe with drink correspinding to those critieria.
}
\examples{
filtering_fun(cocktails, "gin")
}
