% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingredientsTable.R
\name{ingredientsTable}
\alias{ingredientsTable}
\title{Render Ingredient-Quantity Table}
\usage{
ingredientsTable(df = cocktails, selected_cocktail)
}
\arguments{
\item{df}{The cocktails dataframe from the package}

\item{selected_cocktail}{A cocktail name string}
}
\value{
An HTML table with ingredients and corresponding quantities
}
\description{
The \code{ingredientsTable} function takes as input the cocktails dataframe and the
name of a cocktail in that dataframe and returns its ingredients along with quantities.
}
\examples{
ingredientsTable(cocktails, "daiquiri")
}
