% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_ing_table.R
\name{render_ing_table}
\alias{render_ing_table}
\title{Render Ingredient-Quantity Table}
\usage{
render_ing_table(selected_cocktail)
}
\arguments{
\item{selected_cocktail}{A cocktail name string}
}
\value{
An html table with ingredients and corresponding quantities
}
\description{
Render Ingredient-Quantity Table
}
\examples{
render_ing_table("daiquiri")
}
