% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderSideIngredientUI.R
\name{renderSideIngredientUI}
\alias{renderSideIngredientUI}
\title{Dropdown rendering}
\usage{
renderSideIngredientUI(input, output)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}
}
\value{
Dynamically created dropdown menus
}
\description{
This functions dynamically renders the dropdown menus in the Shiny app UI
}
