% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surpriseMe.R
\name{surpriseMe}
\alias{surpriseMe}
\title{Surprise Me}
\usage{
surpriseMe()
}
\value{
A randomly selected cocktail within the \code{cocktails} dataframe
}
\description{
The function can be used to randomly select a cocktail from the dataframe.
The output can the be used in the \code{ingredientsTable} function in order
to get the ingredients and quantities needed for the cocktail.
}
\examples{
#using it on its own
surpriseMe()
#using it with ingredientsTable
ingredientsTable(selected_cocktail = surpriseMe()$Name)
}
