% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/welcomeMessage.R
\name{welcomeMessage}
\alias{welcomeMessage}
\title{Welcome message}
\usage{
welcomeMessage()
}
\value{
A welcome message in a popup in shiny
}
\description{
This internal function is called when the app is loaded and shows the welcome message in a popup.
}
