% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterCocktails.R
\name{filterCocktails}
\alias{filterCocktails}
\title{Filtering cocktails list}
\usage{
filterCocktails(df, alcohol = NULL, ing1 = NULL, ing2 = NULL)
}
\arguments{
\item{df}{A dataframe}

\item{alcohol}{A string}

\item{ing1}{A string}

\item{ing2}{A string}
}
\value{
a filtered drinks dataframe based on inputs
}
\description{
The `cocktail_list` function is used to create the list of
drinks based on inputs from the dropdowns.
It filters the datasets based on on alcohol and two side ingredients
an returns a filtered dataframe with drink corresponding to those criteria.
}
\examples{
cocktailList(cocktails, "gin")
}
