% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateSideIngredients.R
\name{updateSideIngredients}
\alias{updateSideIngredients}
\title{Update side ingredients content}
\usage{
updateSideIngredients(input, session, cocktails)
}
\arguments{
\item{input}{Shiny input vector}

\item{session}{Shiny output vector}

\item{cocktails}{Cocktails dataframe}
}
\value{
Updated dropdown menus
}
\description{
This function dynamically updates the content inside the side ingredient dropdowns
based on what alcohol was selected. It also prevents that a user can choose the
same side ingredient twice.
}
